remove dependency on web-time in solver
