+++ solvers/Cargo.toml
@@ -14,7 +14,6 @@ radix-heap = "0.4.2"
 rustc-hash = "1.1.0"
 pareto_front = "1.0.1"
 bitfield-struct = "0.8.0"
-web-time = "1.1.0"
 
 [dev-dependencies]
 rand = "0.8.5"
